culture-pom-heroku/requirements.txt
streamlit>=1.29.0
pandas>=2.2.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
python-dateutil>=2.8.2
openpyxl>=3.1.2
